transport: locator: add TokenRing struct
New module was created - locator. Its name is inspired by analogous
module in ScyllaDB source tree, in which different replication
strategies live.

Cassandra and Scylla use the token ring for determining data ownership
which allows for efficient load balancing. The token ring is used by the
driver to find the replicas for a given token. Each ring member has a
token (i64 number) which defines the member's position on the ring. The
ring is circular and can be traversed in the order of increasing tokens.
`TokenRing` makes it easy and efficient to traverse the ring starting at
a given token.

Co-authored-by: Michał Sala <michal.sala@scylladb.com>